llmops-rag-gateway/requirements.txt
# =============================================================================
# LLM Gateway — Python dependencies
# Install: pip install -r requirements.txt
# =============================================================================

# --- Web framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# --- Auth ---
PyJWT==2.8.0

# --- Observability ---
prometheus-client==0.20.0

# --- LLM / AI (only needed when LLM_BACKEND=openai) ---
openai>=1.25.0

# --- Vector store (only needed when VECTOR_STORE_BACKEND=pgvector) ---
asyncpg==0.29.0

# --- HTTP client (eval harness) ---
httpx==0.27.0

# --- Testing ---
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# --- Code quality ---
ruff==0.4.2
